+++ tests/roots/test-announcement/conf.py
@@ -1,0 +1,7 @@
+from sphinx_clarity_theme import ThemeOptions
+
+html_theme = "sphinx_clarity_theme"
+
+html_theme_options: ThemeOptions = {
+    "announcement": """<strong>Deprecated version!</strong> See this page in the latest  <a href="/stable/$PAGE$">stable version</a>""",
+}